Move binary methods to its own helper
