+++ src/api/app/helpers/webui/packages/binaries_helper.rb
@@ -1,0 +1,7 @@
+module Webui::Packages::BinariesHelper
+  include Webui::WebuiHelper
+
+  def uploadable?(filename, architecture)
+    ::Cloud::UploadJob.new(filename: filename, arch: architecture).uploadable?
+  end
+end
